feat. 로그아웃 연결, 유저 인증정보에 따라 헤더 상태 변경 기능 추가
