+++ src/api/supabaseAuth.ts
@@ -9,7 +9,7 @@ export class AuthError extends Error {
   }
 }
 
-export const googleLogin = async () => {
+export const signin = async () => {
   const { error } = await auth.signInWithOAuth({
     provider: 'google',
     options: { queryParams: { access_type: 'offline', prompt: 'consent' } },
@@ -17,7 +17,7 @@ export const googleLogin = async () => {
   if (error?.status) throw new AuthError('로그인 정보가 잘못되었습니다.');
 };
 
-export const logout = async () => {
+export const signout = async () => {
   const { error } = await auth.signOut();
 
   if (error?.status) throw new AuthError('로그아웃에 실패하였습니다.');
